🚚 MOVE: main -> shared 로 파일명 변경
